Synchronize with the latest reactor-c
reactor-c commit ID: f64c02d3c0eec91de7d7418bf4c881c5e28b64b1
